Move getParentEnvironment into pty service
Part of #117257
