+++ src/vs/platform/terminal/node/terminalEnvironment.ts
@@ -3,13 +3,92 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import * as fs from 'fs';
 import * as os from 'os';
 import * as path from 'vs/base/common/path';
-import * as process from 'vs/base/common/process';
 import { exists } from 'vs/base/node/pfs';
 import { isString } from 'vs/base/common/types';
 import { getCaseInsensitive } from 'vs/base/common/objects';
-import { IProcessEnvironment, isWindows } from 'vs/base/common/platform';
+import { IProcessEnvironment, isLinux, isMacintosh, isWindows } from 'vs/base/common/platform';
+
+let parentEnv: IProcessEnvironment | undefined;
+
+export async function getParentEnvironment(baseEnvironment: IProcessEnvironment = process.env): Promise<IProcessEnvironment> {
+	if (parentEnv) {
+		return parentEnv as IProcessEnvironment;
+	}
+
+	// For Linux use /proc/<pid>/status to get the parent of the main process and then fetch its
+	// env using /proc/<pid>/environ.
+	if (isLinux) {
+		const mainProcessId = process.ppid;
+		const codeProcessName = path.basename(process.argv[0]);
+		let pid: number = 0;
+		let ppid: number = mainProcessId;
+		let name: string = codeProcessName;
+		do {
+			pid = ppid;
+			const status = await fs.promises.readFile(`/proc/${pid}/status`, 'utf8');
+			const splitByLine = status.split('\n');
+			splitByLine.forEach(line => {
+				if (line.indexOf('Name:') === 0) {
+					name = line.replace(/^Name:\s+/, '');
+				}
+				if (line.indexOf('PPid:') === 0) {
+					ppid = parseInt(line.replace(/^PPid:\s+/, ''));
+				}
+			});
+		} while (name === codeProcessName);
+		const rawEnv = await fs.promises.readFile(`/proc/${pid}/environ`, 'utf8');
+		const env: IProcessEnvironment = {};
+		rawEnv.split('\0').forEach(e => {
+			const i = e.indexOf('=');
+			env[e.substr(0, i)] = e.substr(i + 1);
+		});
+		parentEnv = env;
+		return env;
+	}
+
+	// For macOS we want the "root" environment as shells by default run as login shells. It
+	// doesn't appear to be possible to get the "root" environment as `ps eww -o command` for
+	// PID 1 (the parent of the main process when launched from the dock/finder) returns no
+	// environment, because of this we will fill in the root environment using a allowlist of
+	// environment variables that we have.
+	if (isMacintosh) {
+		const env: IProcessEnvironment = {};
+		// This list was generated by diffing launching a terminal with {} and the system
+		// terminal launched from finder.
+		const rootEnvVars = [
+			'SHELL',
+			'SSH_AUTH_SOCK',
+			'Apple_PubSub_Socket_Render',
+			'XPC_FLAGS',
+			'XPC_SERVICE_NAME',
+			'HOME',
+			'LOGNAME',
+			'TMPDIR'
+		];
+		rootEnvVars.forEach(k => {
+			if (baseEnvironment[k]) {
+				env[k] = baseEnvironment[k]!;
+			}
+		});
+		parentEnv = env;
+		return env;
+	}
+
+	// TODO: Windows should return a fresh environment block, might need native code?
+	if (isWindows) {
+		parentEnv = baseEnvironment;
+	}
+
+	// Unexpected platform, fallback to base
+	if (!parentEnv) {
+		parentEnv = baseEnvironment;
+	}
+
+	return parentEnv as IProcessEnvironment;
+}
 
 export function getWindowsBuildNumber(): number {
 	const osVersion = (/(\d+)\.(\d+)\.(\d+)/g).exec(os.release());
