+++ src/vs/workbench/contrib/terminal/browser/remoteTerminalService.ts
@@ -6,6 +6,7 @@
 import { Emitter } from 'vs/base/common/event';
 import { Disposable } from 'vs/base/common/lifecycle';
 import { revive } from 'vs/base/common/marshalling';
+import { IProcessEnvironment } from 'vs/base/common/platform';
 import { URI } from 'vs/base/common/uri';
 import { localize } from 'vs/nls';
 import { ICommandService } from 'vs/platform/commands/common/commands';
@@ -197,5 +198,13 @@ export class RemoteTerminalService extends Disposable implements IRemoteTerminal
 		}
 
 		return this._remoteTerminalChannel.getTerminalLayoutInfo();
+	}
+
+
+	public getParentEnvironment(): Promise<IProcessEnvironment> {
+		if (!this._remoteTerminalChannel) {
+			throw new Error(`Cannot call getParentEnvironment when there is no remote`);
+		}
+		return this._remoteTerminalChannel.getParentEnvironment();
 	}
 }
