+++ src/vs/workbench/contrib/terminal/common/remoteTerminalChannel.ts
@@ -21,6 +21,7 @@ import { IEnvironmentVariableService, ISerializableEnvironmentVariableCollection
 import { IProcessDataEvent, IShellLaunchConfig, IShellLaunchConfigDto, ITerminalDimensionsOverride, ITerminalEnvironment, ITerminalLaunchError, ITerminalsLayoutInfo, ITerminalsLayoutInfoById, TerminalShellType } from 'vs/platform/terminal/common/terminal';
 import { ITerminalConfiguration, TERMINAL_CONFIG_SECTION } from 'vs/workbench/contrib/terminal/common/terminal';
 import { IGetTerminalLayoutInfoArgs, IProcessDetails, IPtyHostProcessReplayEvent, ISetTerminalLayoutInfoArgs } from 'vs/platform/terminal/common/terminalProcess';
+import { IProcessEnvironment } from 'vs/base/common/platform';
 
 export const REMOTE_TERMINAL_CHANNEL_NAME = 'remoteterminal';
 
@@ -281,5 +282,9 @@ export class RemoteTerminalChannelClient {
 			workspaceId: workspace.id,
 		};
 		return this._channel.call<ITerminalsLayoutInfo>('$getTerminalLayoutInfo', args);
+	}
+
+	public getParentEnvironment(): Promise<IProcessEnvironment> {
+		return this._channel.call<IProcessEnvironment>('$getParentEnvironment');
 	}
 }
